pyproject.toml
[project]
name = "doc-extractor"
version = "0.1.0"
description = "RESTful microservice for document upload, classification and structured data extraction using LLMs."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.2",
    "google-cloud-aiplatform>=1.140.0",
    "httpx>=0.28.1",
    "pydantic-ai[google]>=0.2.0",
    "anthropic[vertex]>=0.52.0",
    "pydantic-settings>=2.12.0",
    "pymupdf>=1.27.2.2",
    "google-auth>=2.49.1",
    "docling>=2.81.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/backend", "experiments"]
